feat: ios: Rust API update
