--- ios/PolkadotVault/Backend/NavigationServices/SettingsTabService.swift
@@ -1,23 +1,0 @@
-//
-//  SettingsTabService.swift
-//  PolkadotVault
-//
-//  Created by Krzysztof Rodak on 08/05/2023.
-//
-
-import Foundation
-
-final class SettingsTabService {
-    private let navigation: NavigationCoordinator
-
-    init(
-        navigation: NavigationCoordinator = NavigationCoordinator()
-    ) {
-        self.navigation = navigation
-    }
-
-    func onSettingsTabDisplay() {
-        navigation.performFake(navigation: .init(action: .start))
-        navigation.performFake(navigation: .init(action: .navbarSettings))
-    }
-}
